requirements.txt
mysql-connector-python==8.0.18
bbcode==1.0.33